chore: resolve all data asynchronously
